backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mongoengine==0.27.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-socketio==5.8.0
aiofiles==23.2.1
Pillow==10.1.0
cryptography==41.0.8
smtplib-ssl==1.0.0
python-multipart==0.0.6
psutil==5.9.6
httpx==0.25.2
opencv-python-headless==4.13.0.92
svgpathtools==1.7.2
PyMuPDF==1.27.1
numpy==2.4.2
scipy==1.17.1
svgwrite==1.4.3
pypotrace==0.1.0
